refactor: move SSV integration tests to proper tests/ directory
- Move integration tests from src/lib.rs to tests/integration.rs following Rust conventions
- Make event_processor and generated modules public for test access
- All 5 integration tests passing in proper integration test location
